MixedDrinkTest: testGetVolumeNonAlcoholic(), MixedDrinkTest: getVolume()
Die Vergleiche in testGetVolumeNonAlcoholic implementiert und im Mixed Drink getVolume die Volume Variable zu einer Double gemacht

+++ Drinks/src/main/java/at/fhj/iit/MixedDrink.java
@@ -28,7 +28,7 @@ public class MixedDrink extends Drink{
      */
     @Override
     public double getVolume() {
-        int volume = 0;
+        double volume = 0;
         for(int i = 0; i < liquidArray.length; i++){
             volume += liquidArray[i].getVolume();
         }
